wedding: initial account creation and validation
Added in the following models:
- profile
- rsvp
- RSVP person

An RSVP is foreign-keyed to by pre-created RSVP people. The RSVP
has a last name and address that it's stored under so that it can
only be claimed by once by (hopefully) the rightful owner.

Create-account form crated to assist in account creation that seems
to do a pretty decent job.

RSVP page built using a standard view and an inline form-factory
to list all of the foreign-keyed RSVP people associated with the
user's RSVP.